move reducer to seperate file
closes #176
